+++ CHANGES.md
@@ -2,6 +2,8 @@ Unreleased
 ----------
 
 * Add `Re.group_count` to get the number of groups in a compiled regex (#218)
+* Add `Re.exec_partial_detailed` to allow resuming searches from partial inputs
+  (#219)
 
 1.10.4 (27-Apr-2022)
 --------------------